refactor(platform-settings): group platform families
