--- src/frontend/features/assistants/AssistantListSearchBar.tsx
@@ -1,2 +1,0 @@
-export { AssistantListSearchBar } from './AssistantListSearchBar.android';
-export type { AssistantListSearchBarProps } from './assistantListSearchBar.types';
