+++ src/frontend/features/assistants/AssistantListSearchBar/AssistantListSearchBar.android.tsx
@@ -1,4 +1,4 @@
-import type { AssistantListSearchBarProps } from './assistantListSearchBar.types';
+import type { AssistantListSearchBarProps } from './AssistantListSearchBar.types';
 
 export function AssistantListSearchBar(_props: AssistantListSearchBarProps) {
   return null;
