+++ src/frontend/features/assistants/AssistantListSearchBar/AssistantListSearchBar.tsx
@@ -1,0 +1,2 @@
+export { AssistantListSearchBar } from './AssistantListSearchBar.android';
+export type { AssistantListSearchBarProps } from './AssistantListSearchBar.types';
