+++ src/frontend/features/assistants/AssistantListSearchBar/AssistantListSearchBar.types.ts
