--- src/frontend/features/settings/DataDirectorySection.tsx
@@ -1,2 +1,0 @@
-export { DataDirectorySection } from './DataDirectorySection.android';
-export type { DataDirectorySectionProps } from './dataDirectorySection.types';
