+++ src/frontend/features/settings/DataDirectorySection/DataDirectorySection.tsx
@@ -1,0 +1,2 @@
+export { DataDirectorySection } from './DataDirectorySection.android';
+export type { DataDirectorySectionProps } from './DataDirectorySection.types';
