+++ src/frontend/features/settings/PermissionsScreen/PermissionListPresentation/PermissionListPresentation.ios.tsx
@@ -1,6 +1,6 @@
 import { Image } from '@/frontend/components/nativePrimitives';
 
-import type { PermissionKind } from './permissionConfig';
+import type { PermissionKind } from '../permissionConfig';
 
 export const visiblePermissionKinds = [
   'location',
@@ -10,10 +10,10 @@ export const visiblePermissionKinds = [
 ] as const satisfies readonly PermissionKind[];
 
 const permissionImages: Record<PermissionKind, number> = {
-  calendar: require('../../../../../assets/permissions/ios/calendar.png'),
-  health: require('../../../../../assets/permissions/ios/health.png'),
-  location: require('../../../../../assets/permissions/ios/location.png'),
-  reminders: require('../../../../../assets/permissions/ios/reminders.png'),
+  calendar: require('../../../../../../assets/permissions/ios/calendar.png'),
+  health: require('../../../../../../assets/permissions/ios/health.png'),
+  location: require('../../../../../../assets/permissions/ios/location.png'),
+  reminders: require('../../../../../../assets/permissions/ios/reminders.png'),
 };
 
 export function PermissionListLeading({ kind }: { kind: PermissionKind }) {
