--- src/frontend/features/settings/ProviderScreen/models/components/ProviderModelListLayout.tsx
@@ -1,2 +1,0 @@
-export { ProviderModelListLayout } from './ProviderModelListLayout.android';
-export type { ProviderModelListLayoutProps } from './providerModelListLayout.types';
