+++ src/frontend/features/settings/ProviderScreen/models/components/ProviderModelListLayout/ProviderModelListLayout.ios.tsx
@@ -1,6 +1,6 @@
-import { ProviderModelListContent } from './ProviderModelListContent';
-import type { ProviderModelListLayoutProps } from './providerModelListLayout.types';
-import { ProviderModelSearchField } from './ProviderModelSearchField';
+import { ProviderModelListContent } from '../ProviderModelListContent';
+import { ProviderModelSearchField } from '../ProviderModelSearchField/ProviderModelSearchField';
+import type { ProviderModelListLayoutProps } from './ProviderModelListLayout.types';
 
 export function ProviderModelListLayout({
   searchText,
