+++ src/frontend/features/settings/ProviderScreen/models/components/ProviderModelListLayout/ProviderModelListLayout.tsx
@@ -1,0 +1,2 @@
+export { ProviderModelListLayout } from './ProviderModelListLayout.android';
+export type { ProviderModelListLayoutProps } from './ProviderModelListLayout.types';
