+++ src/frontend/features/settings/ProviderScreen/models/components/ProviderModelListLayout/ProviderModelListLayout.types.ts
@@ -1,0 +1,10 @@
+import type { ProviderModelListContentProps } from '../ProviderModelListContent';
+import type { ProviderModelSearchFieldProps } from '../ProviderModelSearchField/ProviderModelSearchField.types';
+
+export type ProviderModelListLayoutProps = Omit<
+  ProviderModelListContentProps,
+  'ListHeaderComponent'
+> &
+  ProviderModelSearchFieldProps & {
+    showSearch: boolean;
+  };
