--- src/frontend/features/settings/ProviderScreen/models/components/ProviderModelSearchField.tsx
@@ -1,2 +1,0 @@
-export { ProviderModelSearchField } from './ProviderModelSearchField.android';
-export type { ProviderModelSearchFieldProps } from './providerModelSearchField.types';
