+++ src/frontend/features/settings/ProviderScreen/models/components/ProviderModelSearchField/ProviderModelSearchField.tsx
@@ -1,0 +1,2 @@
+export { ProviderModelSearchField } from './ProviderModelSearchField.android';
+export type { ProviderModelSearchFieldProps } from './ProviderModelSearchField.types';
