--- src/frontend/features/settings/ProviderScreen/models/components/providerModelListLayout.types.ts
@@ -1,10 +1,0 @@
-import type { ProviderModelListContentProps } from './ProviderModelListContent';
-import type { ProviderModelSearchFieldProps } from './providerModelSearchField.types';
-
-export type ProviderModelListLayoutProps = Omit<
-  ProviderModelListContentProps,
-  'ListHeaderComponent'
-> &
-  ProviderModelSearchFieldProps & {
-    showSearch: boolean;
-  };
